Source/Server/requirements.txt
fastapi==0.110.1
paho-mqtt==2.0.0
PyJWT==2.8.0
pymongo==4.7.0
python-dotenv==1.0.1